Rename .mjs to .js
