+++ package.json
@@ -18,10 +18,10 @@
     "build": "nx run-many --target=build --projects=root",
     "watch": "concurrently --kill-others \"pnpm -r build:rollup -w\" \"pnpm -r build:types --watch\"",
     "dev": "pnpm run watch",
-    "prettier": "prettier --plugin-search-dir . \"{packages,examples,scripts}/**/*.{md,js,jsx,cjs,mjs,ts,tsx,json,vue,svelte}\"",
+    "prettier": "prettier --plugin-search-dir . \"{packages,examples,scripts}/**/*.{md,js,jsx,cjs,ts,tsx,json,vue,svelte}\"",
     "prettier:write": "pnpm run prettier --write",
-    "cipublish": "node scripts/publish.mjs",
-    "validatePackages": "node scripts/validate-packages.mjs"
+    "cipublish": "node scripts/publish.js",
+    "validatePackages": "node scripts/validate-packages.js"
   },
   "nx": {
     "includedScripts": [
