+++ packages/query-devtools/rollup.config.js
@@ -1,7 +1,7 @@
 // @ts-check
 
 import { defineConfig } from 'rollup'
-import { buildConfigs } from '../../scripts/getRollupConfig.mjs'
+import { buildConfigs } from '../../scripts/getRollupConfig.js'
 
 export default defineConfig(
   buildConfigs({
