+++ packages/query-sync-storage-persister/rollup.config.js
@@ -1,11 +1,11 @@
 // @ts-check
 
 import { defineConfig } from 'rollup'
-import { buildConfigs } from '../../scripts/getRollupConfig.mjs'
+import { buildConfigs } from '../../scripts/getRollupConfig.js'
 
 export default defineConfig(
   buildConfigs({
-    name: 'query-core',
+    name: 'query-sync-storage-persister',
     outputFile: 'index',
     entryFile: './src/index.ts',
   }),
