+++ packages/react-query-persist-client/rollup.config.js
@@ -1,11 +1,11 @@
 // @ts-check
 
 import { defineConfig } from 'rollup'
-import { buildConfigs } from '../../scripts/getRollupConfig.mjs'
+import { buildConfigs } from '../../scripts/getRollupConfig.js'
 
 export default defineConfig(
   buildConfigs({
-    name: 'query-sync-storage-persister',
+    name: 'react-query-persist-client',
     outputFile: 'index',
     entryFile: './src/index.ts',
   }),
