+++ packages/react-query/package.json
@@ -31,7 +31,7 @@
     "test:lib": "vitest run --coverage",
     "test:lib:dev": "pnpm run test:lib --watch",
     "build": "pnpm build:rollup && pnpm build:codemods && pnpm build:types",
-    "build:rollup": "rollup --config rollup.config.mjs",
+    "build:rollup": "rollup --config rollup.config.js",
     "build:codemods": "cpy ../codemods/src/**/* ./build/codemods",
     "build:types": "tsc --emitDeclarationOnly"
   },
