+++ packages/solid-query/package.json
@@ -39,7 +39,7 @@
     "test:lib": "vitest run --coverage",
     "test:lib:dev": "pnpm run test:lib --watch",
     "build": "pnpm build:rollup && pnpm build:types",
-    "build:rollup": "rollup --config rollup.config.mjs",
+    "build:rollup": "rollup --config rollup.config.js",
     "build:types": "tsc"
   },
   "files": [
