+++ packages/solid-query/tsconfig.eslint.json
@@ -3,5 +3,5 @@
   "compilerOptions": {
     "noEmit": true
   },
-  "include": ["**/*.ts", "**/*.tsx", ".eslintrc.cjs", "rollup.config.mjs"]
+  "include": ["**/*.ts", "**/*.tsx", ".eslintrc.cjs", "rollup.config.js"]
 }
