+++ packages/vue-query/rollup.config.js
@@ -1,11 +1,11 @@
 // @ts-check
 
 import { defineConfig } from 'rollup'
-import { buildConfigs } from '../../scripts/getRollupConfig.mjs'
+import { buildConfigs } from '../../scripts/getRollupConfig.js'
 
 export default defineConfig(
   buildConfigs({
-    name: 'react-query-persist-client',
+    name: 'vue-query',
     outputFile: 'index',
     entryFile: './src/index.ts',
   }),
