+++ scripts/config.js
