+++ scripts/validatePackages.js
@@ -4,7 +4,7 @@ import path from 'node:path'
 import fsp from 'node:fs/promises'
 import jsonfile from 'jsonfile'
 import { publint } from 'publint'
-import { packages, rootDir } from './config.mjs'
+import { packages, rootDir } from './config.js'
 
 async function run() {
   console.info('Validating packages...')
